# packages/backend-nestjs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User Created Successfully"
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Register User",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_forgotPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordDto"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated user password"
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/update-password": {
      "post": {
        "operationId": "AuthController_updatePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated user password"
          },
          "400": {
            "description": "Validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/diary": {
      "get": {
        "operationId": "DiaryController_getAllData",
        "parameters": [
          {
            "name": "date",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JWT_DTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Room created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetAllDiaryDataResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Diary"
        ]
      }
    },
    "/diary/entry/personal": {
      "post": {
        "operationId": "DiaryController_postPersonalEntry",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePersonalNotesResponseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Diary"
        ]
      },
      "put": {
        "operationId": "DiaryController_putPersonalEntry",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePersonalNotesResponseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Diary"
        ]
      },
      "delete": {
        "operationId": "DiaryController_deletePersonalEntry",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeletePersonalEntryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Diary"
        ]
      }
    },
    "/diary/entry/work": {
      "post": {
        "operationId": "DiaryController_postWorkEntry",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkNotesResponseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Diary"
        ]
      },
      "put": {
        "operationId": "DiaryController_putWorkEntry",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkNotesResponseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Diary"
        ]
      },
      "delete": {
        "operationId": "DiaryController_deleteWorkEntry",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteWorkEntryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Diary"
        ]
      }
    },
    "/diary/entry/gym": {
      "post": {
        "operationId": "DiaryController_postGymEntry",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGymNotesResponseDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Diary"
        ]
      },
      "put": {
        "operationId": "DiaryController_putGymEntry",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGymNotesResponseDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Diary"
        ]
      },
      "delete": {
        "operationId": "DiaryController_deleteGymEntry",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteGymEntryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Diary"
        ]
      }
    },
...
  },
  "info": {
    "title": "Vipul Waghmare APIs",
    "description": "All the APIs created",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "example": "john.doe@email.com"
          },
          "password": {
            "type": "string",
            "example": "password@1234"
          }
        },
        "required": [
          "name",
          "email",
          "password"
        ]
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "example": "field_name",
            "description": "Field that caused the error"
          },
          "message": {
            "type": "string",
            "example": "Field is required",
            "description": "Error message"
          }
        },
        "required": [
          "field",
          "message"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400,
            "description": "HTTP status code"
          },
          "error": {
            "type": "string",
            "example": "Bad Request",
            "description": "Error type"
          },
          "details": {
            "description": "List of validation errors",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        },
        "required": [
          "statusCode",
          "error",
          "details"
        ]
      },
      "LoginUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john.doe@email.com"
          },
          "password": {
            "type": "string",
            "example": "password@1234"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "string",
            "example": "User is logged in with email: user@example.com",
            "description": "Success message"
          },
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "description": "JWT access token"
          },
          "userId": {
            "type": "string",
            "example": "507f1f77bcf86cd799439011",
            "description": "MongoDB user ID"
          },
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "User email"
          },
          "name": {
            "type": "string",
            "example": "John Doe",
            "description": "User name"
          }
        },
        "required": [
          "success",
          "accessToken",
          "userId",
          "email",
          "name"
        ]
      },
      "ForgotPasswordDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "john.doe@email.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "token-received-in-mail"
          },
          "password": {
            "type": "string",
            "example": "new.password@1234"
          }
        },
        "required": [
          "token",
          "password"
        ]
      },
      "UpdatePasswordDto": {
        "type": "object",
        "properties": {
          "newPassword": {
            "type": "string",
            "example": "new.password@1234"
          },
          "password": {
            "type": "string",
            "example": "password@1234"
          }
        },
        "required": [
          "newPassword",
          "password"
        ]
      },
...
      "JWT_DTO": {
        "type": "object",
        "properties": {}
      },
...
      "GetAllDiaryDataResponseDto": {
        "type": "object",
        "properties": {
          "personal": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetPersonalNotesResponseDto"
            }
          },
          "work": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetWorkNotesResponseDto"
            }
          },
          "gym": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GetGymProgressResponseDto"
            }
          },
          "health": {
            "$ref": "#/components/schemas/GetHealthLogResponseDto"
          },
          "weight": {
            "type": "number",
            "example": 11,
            "description": "The weight of the person in kg"
          },
          "height": {
            "type": "number",
            "example": 12,
            "description": "The height of the person in cm"
          },
          "_id": {
            "type": "string",
            "example": "67eb8c8028300395ab7df7a0",
            "description": "id of the event"
          },
          "userId": {
            "type": "string",
            "example": "67eb8c8028300395ab7df7a0",
            "description": "user id"
          },
          "date": {
            "type": "string",
            "example": "2025-03-26",
            "description": "date"
          }
        },
        "required": [
          "personal",
          "work",
          "gym",
          "health",
          "weight",
          "height",
          "_id",
          "userId",
          "date"
        ]
      },
      "CreatePersonalNotesResponseDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Weekend Trip to the Mountains",
            "description": "The title of the event"
          },
          "content": {
            "type": "string",
            "example": "Had an amazing time hiking in the mountains this weekend",
            "description": "A brief description of the event"
          },
          "date": {
            "type": "string",
            "example": "2025-03-26",
            "description": "The date of the event"
          }
        },
        "required": [
          "title",
          "content",
          "date"
        ]
      },
      "CreateWorkNotesResponseDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Client Meeting",
            "description": "The title of the event"
          },
          "content": {
            "type": "string",
            "example": "Had a productive meeting with the client.",
            "description": "A brief description of the event"
          },
          "date": {
            "type": "string",
            "example": "2025-03-26",
            "description": "The date of the event"
          }
        },
        "required": [
          "title",
          "content",
          "date"
        ]
      },
...
      "CreateGymNotesResponseDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "Strength Training",
            "description": "The type of workout"
          },
          "duration": {
            "type": "number",
            "example": 60,
            "description": "The duration of the workout in minutes"
          },
          "exercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GymProgressExerciseDto"
            }
          },
          "notes": {
            "type": "string",
            "example": "Great workout today. Increased weight on bench press.",
            "description": "A brief description of the workout"
          },
          "date": {
            "type": "string",
            "example": "2025-03-26",
            "description": "The date of the event"
          }
        },
        "required": [
          "type",
          "duration",
          "exercises",
          "notes",
          "date"
        ]
      },
...
      "UpdatePersonalNotesResponseDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Weekend Trip to the Mountains",
            "description": "The title of the event"
          },
          "content": {
            "type": "string",
            "example": "Had an amazing time hiking in the mountains this weekend",
            "description": "A brief description of the event"
          },
          "date": {
            "type": "string",
            "example": "2025-03-26",
            "description": "The date of the event"
          },
          "id": {
            "type": "string",
            "example": "12345678-1234-1234-1234-123456789012",
            "description": "The ID of the event"
          }
        },
        "required": [
          "title",
          "content",
          "date",
          "id"
        ]
      },
      "UpdateWorkNotesResponseDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "Client Meeting",
            "description": "The title of the event"
          },
          "content": {
            "type": "string",
            "example": "Had a productive meeting with the client.",
            "description": "A brief description of the event"
          },
          "date": {
            "type": "string",
            "example": "2025-03-26",
            "description": "The date of the event"
          },
          "id": {
            "type": "string",
            "example": "12345678-1234-1234-1234-123456789012",
            "description": "The ID of the event"
          }
        },
        "required": [
          "title",
          "content",
          "date",
          "id"
        ]
      },
      "UpdateGymNotesResponseDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "Strength Training",
            "description": "The type of workout"
          },
          "duration": {
            "type": "number",
            "example": 60,
            "description": "The duration of the workout in minutes"
          },
          "exercises": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GymProgressExerciseDto"
            }
          },
          "notes": {
            "type": "string",
            "example": "Great workout today. Increased weight on bench press.",
            "description": "A brief description of the workout"
          },
          "date": {
            "type": "string",
            "example": "2025-03-26",
            "description": "The date of the event"
          },
          "id": {
            "type": "string",
            "example": "12345678-1234-1234-1234-123456789012",
            "description": "The ID of the event"
          }
        },
        "required": [
          "type",
          "duration",
          "exercises",
          "notes",
          "date",
          "id"
        ]
      },
...
      "DeletePersonalEntryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "date": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "date"
        ]
      },
      "DeleteWorkEntryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "date": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "date"
        ]
      },
      "DeleteGymEntryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "date": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "date"
        ]
      },
...
    }
  }
}